The rulestest was really deeply nested - this factors that repetitious, deeply nested code out into repetitious, less-nested code. I still need to work on getting rid of the repetition.

+++ _ceu_app.h
@@ -14,7 +14,7 @@
 
 /* TODO: lbl => unsigned */
 #ifndef CEU_OS
-typedef s8 tceu_nlbl;
+typedef s16 tceu_nlbl;
 #endif
 
 #ifdef CEU_IFCS
